fix: explain Linux service inspection failures
Preserve safe user-bus and busctl diagnostics through Doctor and update
without relaxing effective-command inspection or service ownership.
Distinguish a missing bus socket from a missing systemctl executable so
cleanup cannot remove a unit after an unverified disable.

Keep native output and malformed manager JSON out of safe diagnostics.
Isolate the affected service fixtures and split Doctor maintenance tests
without dropping their existing behavior coverage.

Fixes #137503. Reported by @MattFaus.
